Harden repo client disposal concurrency and split setup tests
